asrock/e3c246d4i: Add Intel Coffee Lake board
This board has upstream OpenBMC support so it's quite nice to develop
on (with some extra patches to enable access to the host flash).

This board very loosely used the prodrive/hermes and system76/oryp5 as
a base and inspiration.

What is tested:
- ast2500 BMC: video, serial
- 10G NIC
- USB boot: from virtual CD from BMC and real disk
- EDK2 with UefiPayloadPkg
- 8G DIMM in slot0

What does not work:
- SeaBIOS: seems to hang in the menu. Possibly related to running
  Option ROMs of the 10G NICs.

TODO: A lot of things are pretty bare in this port like USB and PCIe
setup, but I don't own schematics so it's hard to improve on that.
No idea what the nuvoton superio is hooked up to either.

Change-Id: I66e168c8af5de9862b0724fa397ecd709843af1a
Signed-off-by: Arthur Heymans <arthur@aheymans.xyz>